feat: Add Udacity's feature_format.py in a tools folder
The code 'feature_format.py' provided by Udacity was added inside a
tools folder to be used for additional future tools. feature_format.py
contains the functions 'featureFormat' and 'targetFeatureSplit'.
'featureFormat' converts a dictionary to a numpy array of features.
'targetFeatureSplit' splits this array into a list of target values
and a list of feature values for use in ML training.