refactor: Remove loop of ML algorithm, replace with dump to tester.py
Intstead of looping the feature selection, fitting, and classifying with
different random states and taking an average of the accuracy, precision,
and recall values.  I dump the classifier, selected features, and dataset
to files to be used by tester.py (which makes use of StratifiedShuffleSplit
to obtain the accuracy, precision, and recall values.